created actions folder for async actions and to split code.. moved get pizzas to pizza-actions.ts
